control de sesion usuariocontroller

+++ Controllers/UsuarioController.cs
@@ -18,37 +18,84 @@ public class UsuarioController : Controller
     [HttpGet]
     public IActionResult Index()
     {
-        var usuarios = usuarioRepository.GetAllUsuarios();
-        return View(usuarios);
+        if(HttpContext.Session.GetString("Rol")==null){
+            return RedirectToRoute(new{controller = "Login", action = "Index"});
+        }else if(isAdmin()){
+            List<Usuario> usuarios = usuarioRepository.GetAllUsuarios();
+            return View(usuarios);
+        }else{
+            List<Usuario> usuarios = usuarioRepository.GetAllUsuarios().FindAll(u => u.Id ==  HttpContext.Session.GetInt32("id"));
+            return View(usuarios);
+        }
     }
 
     [HttpGet]
     public IActionResult CrearUsuario()
     {
-        return View(new Usuario());
+        if(isAdmin()){
+            return View(new Usuario());
+        }
+        return RedirectToRoute(new{controller = "Login", action = "Index"});
     }
     [HttpPost]
     public IActionResult CrearUsuario(Usuario usuario)
     {
-        usuarioRepository.AddUsuario(usuario);
-        return RedirectToAction("Index");
+        if(isAdmin()){
+            usuarioRepository.AddUsuario(usuario);
+            return RedirectToAction("Index");
+        }
+        return RedirectToRoute(new{controller = "Login", action = "Index"});
     }
     [HttpGet]
     public IActionResult ModificarUsuario(int id)
-    {
-        var usuario = usuarioRepository.GetAllUsuarios().FirstOrDefault(u => u.Id == id);
-        return View(usuario);
+    {   
+        if(HttpContext.Session.GetString("Rol")==null){
+            return RedirectToRoute(new{controller = "Login", action = "Index"});
+        }else if(isAdmin()){
+            var usuario = usuarioRepository.GetAllUsuarios().FirstOrDefault(u => u.Id == id);
+            return View(usuario);
+        }else{
+            if(HttpContext.Session.GetInt32("id")==id){
+                var usuario = usuarioRepository.GetAllUsuarios().FirstOrDefault(u => u.Id == id);
+                return View("ModificarUsuarioOperador",usuario);
+                //return RedirectToAction("ModificarUsuario",usuario);
+            }else{
+                return RedirectToAction("Index");
+            }
+        }
     }
     [HttpPost]
     public IActionResult ModificarUsuario(int id, Usuario usuario)
     {
-        usuarioRepository.UpdateUsuario(id,usuario);
+        if(HttpContext.Session.GetString("Rol")==null){
+            return RedirectToRoute(new{controller = "Login", action = "Index"});
+        }else if(isAdmin()){
+            usuarioRepository.UpdateUsuario(id,usuario);
+        }else{
+            if(HttpContext.Session.GetInt32("id")==id){
+                usuarioRepository.UpdateUsuario(id,usuario);
+            }
+        }
         return RedirectToAction("Index");
     }
     public IActionResult EliminarUsuario(int idUsuario)
     {
-        usuarioRepository.DeleteUsuario(idUsuario);
+        if(HttpContext.Session.GetString("Rol")==null){
+            return RedirectToRoute(new{controller = "Login", action = "Index"});
+        }else if(isAdmin()){
+            usuarioRepository.DeleteUsuario(idUsuario);
+        }else{
+            if(HttpContext.Session.GetInt32("id")==idUsuario){
+                usuarioRepository.DeleteUsuario(idUsuario);
+            }
+        }
         return RedirectToAction("Index");
+    }
+    private bool isAdmin(){
+        if(HttpContext.Session != null && HttpContext.Session.GetString("Rol") == "Administrador"){
+            return true;
+        }
+        return false;
     }
 
     [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
@@ -56,4 +103,5 @@ public class UsuarioController : Controller
     {
         return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
     }
+    
 }
